Refactory completed
