+++ lib/Plugins/Features/AbstractRevisionFeature.php
@@ -15,7 +15,10 @@ use Features;
 use Features\ProjectCompletion\CompletionEventStruct;
 use Features\ReviewExtended\IChunkReviewModel;
 use Features\ReviewExtended\Model\ArchivedQualityReportModel;
+use Features\ReviewExtended\BatchReviewProcessor;
 use Features\ReviewExtended\Model\QualityReportModel;
+use Features\TranslationVersions\Handlers\TranslationEventsHandler;
+use Features\TranslationVersions\Model\TranslationEvent;
 use FilesStorage\AbstractFilesStorage;
 use FilesStorage\FilesStorageFactory;
 use INIT;
@@ -24,12 +27,10 @@ use Log;
 use LQA\ChunkReviewDao;
 use LQA\ChunkReviewStruct;
 use LQA\ModelDao;
-use LQA\ModelStruct;
 use Projects_ProjectDao;
 use Projects_ProjectStruct;
 use Revise_FeedbackDAO;
 use RevisionFactory;
-use SegmentTranslationChangeVector;
 use Utils;
 use ZipArchive;
 
@@ -330,13 +331,13 @@ abstract class AbstractRevisionFeature extends BaseFeature {
     }
 
     /**
-     * @param TranslationVersions\Model\BatchEventCreator $eventCreator
+     * @param TranslationEventsHandler $eventCreator
      *
      * @throws Exception
-     * @internal param SegmentTranslationEventModel $event
+     * @internal param TranslationEvent $event
      */
-    public function batchEventCreationSaved( Features\TranslationVersions\Model\BatchEventCreator $eventCreator ) {
-        $batchReviewProcessor = new Features\ReviewExtended\Model\BatchReviewProcessor( $eventCreator );
+    public function translationVersionSaved( TranslationEventsHandler $eventCreator ) {
+        $batchReviewProcessor = new BatchReviewProcessor( $eventCreator );
         $batchReviewProcessor->process();
     }
 
@@ -557,13 +558,13 @@ abstract class AbstractRevisionFeature extends BaseFeature {
     }
 
     /**
-     * @param SegmentTranslationChangeVector $translation
+     * @param TranslationEvent    $translation
      *
-     * @param ChunkReviewStruct[]            $chunkReviews
+     * @param ChunkReviewStruct[] $chunkReviews
      *
      * @return ISegmentTranslationModel
      */
-    public function getSegmentTranslationModel( SegmentTranslationChangeVector $translation, array $chunkReviews = [] ) {
+    public function getSegmentTranslationModel( TranslationEvent $translation, array $chunkReviews = [] ) {
         $class_name = get_class( $this ) . '\SegmentTranslationModel';
 
         return new $class_name( $translation, $chunkReviews );
