+++ lib/Plugins/Features/ReviewExtended/ISegmentTranslationModel.php
@@ -7,7 +7,7 @@
  */
 
 
-namespace Features;
+namespace Features\ReviewExtended;
 
 interface ISegmentTranslationModel {
     public function getChunkReviewTransitionModel();
