+++ lib/Plugins/Features/ReviewExtended/SegmentTranslationModel.php
@@ -8,34 +8,29 @@
 
 namespace Features\ReviewExtended;
 
-use ChunkReviewTransition\ChunkReviewTransitionModel;
-use ChunkReviewTransition_ChunkReviewTransitionModel;
-use ChunkReviewTransition_UnitOfWork;
 use Chunks_ChunkStruct;
 use Constants;
-use Features\ISegmentTranslationModel;
 use Features\SecondPassReview\Email\RevisionChangedNotificationEmail;
-use Features\SecondPassReview\Model\SegmentTranslationEventDao;
-use Features\TranslationVersions\Model\SegmentTranslationEventStruct;
+use Features\SecondPassReview\Model\TranslationEventDao;
+use Features\TranslationVersions\Model\TranslationEvent;
+use Features\TranslationVersions\Model\TranslationEventStruct;
 use LQA\ChunkReviewStruct;
 use LQA\EntryCommentStruct;
 use LQA\EntryDao;
 use LQA\EntryStruct;
 use LQA\EntryWithCategoryStruct;
 use Routes;
-use SegmentTranslationChangeVector;
 use TransactionableTrait;
 use Users_UserDao;
-use Users_UserStruct;
 
 class SegmentTranslationModel implements ISegmentTranslationModel {
 
     use TransactionableTrait;
 
     /**
-     * @var SegmentTranslationChangeVector
+     * @var TranslationEvent
      */
-    protected $_model;
+    protected $_event;
 
     /**
      * @var Chunks_ChunkStruct
@@ -69,23 +64,23 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
      */
     private $_finalRevisions;
 
-    public function __construct( SegmentTranslationChangeVector $model, array $chunkReviews ) {
-        $this->_model        = $model;
+    public function __construct( TranslationEvent $model, array $chunkReviews ) {
+        $this->_event        = $model;
         $this->_chunkReviews = $chunkReviews;
         $this->_chunk        = $model->getChunk();
         $this->_project      = $this->_chunk->getProject();
 
-        $this->_finalRevisions = ( new SegmentTranslationEventDao() )->getFinalRevisionsForSegment(
-                $this->_chunk->id, $this->_model->getSegmentStruct()->id
+        $this->_finalRevisions = ( new TranslationEventDao() )->getFinalRevisionsForSegment(
+                $this->_chunk->id, $this->_event->getSegmentStruct()->id
         );
 
-        $this->_sourcePagesWithFinalRevisions = array_map( function ( SegmentTranslationEventStruct $event ) {
+        $this->_sourcePagesWithFinalRevisions = array_map( function ( TranslationEventStruct $event ) {
             return $event->source_page;
         }, $this->_finalRevisions );
     }
 
     /**
-     * @return ChunkReviewTransitionModel
+     * @return ChunkReviewTranslationEventTransition
      * @throws \Exception
      */
     public function getChunkReviewTransitionModel() {
@@ -112,10 +107,10 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
 
         $chunkReviews          = [];
         $unsetFinalRevision    = [];
-        $originSourcePage      = $this->_model->getEventModel()->getOriginSourcePage();
-        $destinationSourcePage = $this->_model->getEventModel()->getDestinationSourcePage();
+        $originSourcePage      = $this->_event->getOriginSourcePage();
+        $destinationSourcePage = $this->_event->getDestinationSourcePage();
 
-        $reviewTransitionModel = new ChunkReviewTransitionModel( $this->_model );
+        $segmentReviewTransitionModel = new ChunkReviewTranslationEventTransition( $this->_event );
 
         // populate structs for current segment and propagations
         for ( $i = 0; $i < count( $this->_chunkReviews ); $i++ ) {
@@ -128,11 +123,11 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
             $chunkReview->password    = $this->_chunkReviews[ $i ]->password;
             $chunkReview->source_page = $this->_chunkReviews[ $i ]->source_page;
 
-            if ( $this->_model->isEnteringReviewedState() && $destinationSourcePage == $chunkReview->source_page ) {
+            if ( $this->_event->isEnteringReviewedState() && $destinationSourcePage == $chunkReview->source_page ) {
                 // expect the first chunk review record to be the final
                 // add revised words and advancement
-                $chunkReview->reviewed_words_count += $this->_model->getSegmentStruct()->raw_word_count;
-                $chunkReview->total_tte            += $this->_model->getEventModel()->getCurrentEvent()->time_to_edit;
+                $chunkReview->reviewed_words_count += $this->_event->getSegmentStruct()->raw_word_count;
+                $chunkReview->total_tte            += $this->_event->getCurrentEvent()->time_to_edit;
 
                 if ( $destinationSourcePage != $originSourcePage ) {
                     $chunkReview->advancement_wc += $this->advancementWordCount();
@@ -141,10 +136,10 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
                 $chunkReviews[] = $chunkReview;
 
                 break;
-            } elseif ( $this->_model->isExitingReviewedState() ) {
+            } elseif ( $this->_event->isExitingReviewedState() ) {
                 // expect the direction to be downwards from R3 -> R2 -> R1 etc.
                 if ( in_array( $chunkReview->source_page, $this->_sourcePagesWithFinalRevisions ) ) {
-                    $chunkReview->reviewed_words_count -= $this->_model->getSegmentStruct()->raw_word_count;
+                    $chunkReview->reviewed_words_count -= $this->_event->getSegmentStruct()->raw_word_count;
                     $this->_addIssuesToDelete( $chunkReview->source_page );
                     $chunkReview->penalty_points -= $this->getPenaltyPointsForSourcePage( $chunkReview->source_page );
                     $unsetFinalRevision []       = $chunkReview->source_page;
@@ -156,11 +151,11 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
                     $chunkReviews[]              = $chunkReview;
                 }
             } // TODO: in the following two cases we should consider if the segment is changed or not.
-            elseif ( $this->_model->isBeingLowerReviewed() ) {
+            elseif ( $this->_event->isBeingLowerReviewed() ) {
                 // whenever a revision is lower reviewed we expect the upper revisions to be invalidated.
                 // the reviewed words count is removed from the upper one and moved to the lower one.
                 if ( $originSourcePage == $chunkReview->source_page ) {
-                    $chunkReview->reviewed_words_count -= $this->_model->getSegmentStruct()->raw_word_count;
+                    $chunkReview->reviewed_words_count -= $this->_event->getSegmentStruct()->raw_word_count;
                     $this->_addIssuesToDelete( $chunkReview->source_page );
                     $chunkReview->penalty_points -= $this->getPenaltyPointsForSourcePage( $chunkReview->source_page );
                     $chunkReview->advancement_wc -= $this->advancementWordCount();
@@ -173,23 +168,23 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
                     // TODO: evaluate the case in which the destination revision never received a revision before
                     // evaluate $sourcePagesWithFinalRevisions
                     if ( !in_array( $chunkReview->source_page, $this->_sourcePagesWithFinalRevisions ) ) {
-                        $chunkReview->reviewed_words_count += $this->_model->getSegmentStruct()->raw_word_count;
+                        $chunkReview->reviewed_words_count += $this->_event->getSegmentStruct()->raw_word_count;
                     }
 
                     $chunkReview->advancement_wc += $this->advancementWordCount();
-                    $chunkReview->total_tte      += $this->_model->getEventModel()->getCurrentEvent()->time_to_edit;
+                    $chunkReview->total_tte      += $this->_event->getCurrentEvent()->time_to_edit;
                     $chunkReviews[]              = $chunkReview;
 
                 } elseif ( in_array( $chunkReview->source_page, $this->_sourcePagesWithFinalRevisions ) ) {
                     // this case fits any other intermediate chunkReview record
-                    $chunkReview->reviewed_words_count -= $this->_model->getSegmentStruct()->raw_word_count;
+                    $chunkReview->reviewed_words_count -= $this->_event->getSegmentStruct()->raw_word_count;
                     $this->_addIssuesToDelete( $chunkReview->source_page );
                     $chunkReview->penalty_points -= $this->getPenaltyPointsForSourcePage( $chunkReview->source_page );
                     $unsetFinalRevision[]        = $chunkReview->source_page;
                     $chunkReviews[]              = $chunkReview;
                 }
 
-            } elseif ( $this->_model->isBeingUpperReviewed() ) {
+            } elseif ( $this->_event->isBeingUpperReviewed() ) {
                 if ( $originSourcePage == $chunkReview->source_page ) {
                     // TODO: decide wether or not to remove the revised words
                     // expect advancement to be assigned to the origin source_page
@@ -199,9 +194,9 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
 
                 } elseif ( $destinationSourcePage == $chunkReview->source_page ) {
                     // we reached the last record, destination record of the lower revision, add the count
-                    $chunkReview->reviewed_words_count += $this->_model->getSegmentStruct()->raw_word_count;
+                    $chunkReview->reviewed_words_count += $this->_event->getSegmentStruct()->raw_word_count;
                     $chunkReview->advancement_wc       += $this->advancementWordCount();
-                    $chunkReview->total_tte            += $this->_model->getEventModel()->getCurrentEvent()->time_to_edit;
+                    $chunkReview->total_tte            += $this->_event->getCurrentEvent()->time_to_edit;
                     $chunkReviews[]                    = $chunkReview;
 
                 } elseif ( in_array( $chunkReview->source_page, $this->_sourcePagesWithFinalRevisions ) ) {
@@ -212,8 +207,8 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
                     false;
                 }
             } elseif (
-                    $this->_model->isModifyingICEFromTranslation() &&
-                    $this->_model->getEventModel()->getDestinationSourcePage() == Constants::SOURCE_PAGE_TRANSLATE
+                    $this->_event->isModifyingICEFromTranslation() &&
+                    $this->_event->getDestinationSourcePage() == Constants::SOURCE_PAGE_TRANSLATE
             ) {
                 /**
                  * Enter this condition when we are just changing source page. This change only affects advancement wc.
@@ -222,11 +217,11 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
                 $chunkReview->advancement_wc -= $this->advancementWordCount();
                 $chunkReviews[]              = $chunkReview;
 
-            } elseif ( $this->_model->isEditingCurrentRevision() && $destinationSourcePage == $chunkReview->source_page ) {
-                $chunkReview->total_tte += $this->_model->getEventModel()->getCurrentEvent()->time_to_edit;
+            } elseif ( $this->_event->isEditingCurrentRevision() && $destinationSourcePage == $chunkReview->source_page ) {
+                $chunkReview->total_tte += $this->_event->getCurrentEvent()->time_to_edit;
 
-                if ( $this->_model->isModifyingICEFromRevisionOne() ) {
-                    $chunkReview->reviewed_words_count += $this->_model->getSegmentStruct()->raw_word_count;
+                if ( $this->_event->isModifyingICEFromRevisionOne() ) {
+                    $chunkReview->reviewed_words_count += $this->_event->getSegmentStruct()->raw_word_count;
                 }
 
                 $chunkReviews[] = $chunkReview;
@@ -234,25 +229,25 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
         }
 
         foreach ( $chunkReviews as $chunkReview ) {
-            $reviewTransitionModel->addChunkReview( $chunkReview );
+            $segmentReviewTransitionModel->addChunkReview( $chunkReview );
         }
 
         foreach ( $this->_issuesDeletionList as $issuesToDelete ) {
             foreach ( $issuesToDelete as $issueToDelete ) {
-                $reviewTransitionModel->addIssueToDelete( $issueToDelete );
+                $segmentReviewTransitionModel->addIssueToDelete( $issueToDelete );
             }
         }
 
-        $reviewTransitionModel->setUnsetFinalRevision( $unsetFinalRevision );
+        $segmentReviewTransitionModel->setUnsetFinalRevision( $unsetFinalRevision );
 
-        return $reviewTransitionModel;
+        return $segmentReviewTransitionModel;
     }
 
     /**
      * @throws \Exception
      */
     public function sendNotificationEmail() {
-        if ( $this->_model->getEventModel()->isPropagationSource() && $this->_model->isBeingLowerReviewedOrTranslated() ) {
+        if ( $this->_event->isPropagationSource() && $this->_event->isBeingLowerReviewedOrTranslated() ) {
             $chunkReviewsWithFinalRevisions = [];
             foreach ( $this->_chunkReviews as $chunkReview ) {
                 if ( in_array( $chunkReview->source_page, $this->_sourcePagesWithFinalRevisions ) ) {
@@ -268,7 +263,7 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
      * @param $source_page
      */
     protected function _addIssuesToDelete( $source_page ) {
-        $issue = EntryDao::findByIdSegmentAndSourcePage( $this->_model->getSegmentStruct()->id, $this->_chunk->id, $source_page );
+        $issue = EntryDao::findByIdSegmentAndSourcePage( $this->_event->getSegmentStruct()->id, $this->_chunk->id, $source_page );
 
         if ( $issue ) {
             $this->_issuesDeletionList[ $source_page ] = $issue;
@@ -288,14 +283,14 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
      */
     protected function _sendNotificationEmail( $finalRevisions, $chunkReviewsWithFinalRevisions ) {
         $emails                   = [];
-        $userWhoChangedTheSegment = $this->_model->getEventUser();
-        $revision                 = $chunkReviewsWithFinalRevisions[ $this->_model->getEventModel()->getOriginSourcePage() ];
+        $userWhoChangedTheSegment = $this->_event->getEventUser();
+        $revision                 = $chunkReviewsWithFinalRevisions[ $this->_event->getOriginSourcePage() ];
 
         $serialized_issues = [];
-        if( isset( $this->_issuesDeletionList[ $this->_model->getEventModel()->getOriginSourcePage() ] ) ){
+        if( isset( $this->_issuesDeletionList[ $this->_event->getOriginSourcePage() ] ) ){
 
             /** @var EntryWithCategoryStruct $issue */
-            foreach ( $this->_issuesDeletionList[ $this->_model->getEventModel()->getOriginSourcePage() ] as $k => $issue ) {
+            foreach ( $this->_issuesDeletionList[ $this->_event->getOriginSourcePage() ] as $k => $issue ) {
                 $serialized               = $issue->toArray();
                 $serialized[ 'comments' ] = [];
 
@@ -310,14 +305,14 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
         }
 
         $segmentInfo = [
-                'segment_source'  => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_model->getSegmentStruct()->segment ),
-                'old_translation' => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_model->getEventModel()->getOldTranslation()->translation ),
-                'new_translation' => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_model->getEventModel()->getTranslation()->translation ),
+                'segment_source'  => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_event->getSegmentStruct()->segment ),
+                'old_translation' => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_event->getOldTranslation()->translation ),
+                'new_translation' => \Utils::htmlentitiesToUft8WithoutDoubleEncoding( $this->_event->getTranslation()->translation ),
                 'issues'          => $serialized_issues
         ];
 
         foreach ( $finalRevisions as $finalRevision ) {
-            if ( $finalRevision->source_page != $this->_model->getEventModel()->getOriginSourcePage() ) {
+            if ( $finalRevision->source_page != $this->_event->getOriginSourcePage() ) {
                 continue;
             }
 
@@ -350,7 +345,7 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
         $url    = Routes::revise( $this->_chunk->getProject()->name, $revision->id_job, $revision->review_password,
                 $this->_chunk->source, $this->_chunk->target, [
                         'revision_number' => ReviewUtils::sourcePageToRevisionNumber( $revision->source_page ),
-                        'id_segment'      => $this->_model->getSegmentStruct()->id
+                        'id_segment'      => $this->_event->getSegmentStruct()->id
                 ] );
 
 
@@ -367,11 +362,11 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
     }
 
     protected function advancementWordCount() {
-        if ( $this->_model->getEventModel()->getOldTranslation()->isICE() || $this->_model->getEventModel()->getOldTranslation()->isPreTranslated() ) {
-            return $this->_model->getSegmentStruct()->raw_word_count;
+        if ( $this->_event->getOldTranslation()->isICE() || $this->_event->getOldTranslation()->isPreTranslated() ) {
+            return $this->_event->getSegmentStruct()->raw_word_count;
         }
 
-        return $this->_model->getOldTranslation()->eq_word_count;
+        return $this->_event->getOldTranslation()->eq_word_count;
     }
 
     /**
@@ -399,7 +394,7 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
      * @throws \Exception
      */
     protected function updateFinalRevisionFlag( $unsetFinalRevision ) {
-        $eventStruct = $this->_model->getEventModel()->getCurrentEvent();
+        $eventStruct = $this->_event->getCurrentEvent();
         $is_revision = (int)$eventStruct->source_page > Constants::SOURCE_PAGE_TRANSLATE;
 
         if ( $is_revision ) {
@@ -407,13 +402,13 @@ class SegmentTranslationModel implements ISegmentTranslationModel {
         }
 
         if ( !empty( $unsetFinalRevision ) ) {
-            ( new SegmentTranslationEventDao() )->unsetFinalRevisionFlag(
-                    $this->_chunk->id, [ $this->_model->getSegmentStruct()->id ], $unsetFinalRevision
+            ( new TranslationEventDao() )->unsetFinalRevisionFlag(
+                    $this->_chunk->id, [ $this->_event->getSegmentStruct()->id ], $unsetFinalRevision
             );
         }
 
         $eventStruct->final_revision = $is_revision;
-        SegmentTranslationEventDao::updateStruct( $eventStruct, [ 'fields' => [ 'final_revision' ] ] );
+        TranslationEventDao::updateStruct( $eventStruct, [ 'fields' => [ 'final_revision' ] ] );
     }
 
 }
