+++ lib/Plugins/Features/TranslationVersions.php
@@ -5,13 +5,15 @@ namespace Features;
 use Chunks_ChunkStruct;
 use Exceptions\ControllerReturnException;
 use Exceptions\ValidationError;
-use Features\TranslationVersions\EmptySegmentTranslationVersionHandler;
-use Features\TranslationVersions\Model\BatchEventCreator;
-use Features\TranslationVersions\Model\SegmentTranslationEventModel;
-use Features\TranslationVersions\SegmentTranslationVersionHandler;
+use Features\TranslationVersions\Model\TranslationEvent;
+use Features\TranslationVersions\Handlers\DummyTranslationVersionHandler;
+use Features\TranslationVersions\Handlers\TranslationEventsHandler;
+use Features\TranslationVersions\Handlers\TranslationVersionsHandler;
+use FeatureSet;
 use Jobs_JobDao;
 use Projects_ProjectDao;
 use Projects_ProjectStruct;
+use Translations_SegmentTranslationStruct;
 use Users_UserStruct;
 
 class TranslationVersions extends BaseFeature {
@@ -21,7 +23,7 @@ class TranslationVersions extends BaseFeature {
     public static function getVersionHandlerNewInstance( Chunks_ChunkStruct $chunkStruct, $id_segment, Users_UserStruct $userStruct, Projects_ProjectStruct $projectStruct ) {
 
         if ( $projectStruct->isFeatureEnabled( self::FEATURE_CODE ) ) {
-            return new SegmentTranslationVersionHandler(
+            return new TranslationVersionsHandler(
                     $chunkStruct,
                     $id_segment,
                     $userStruct,
@@ -29,7 +31,7 @@ class TranslationVersions extends BaseFeature {
             );
         }
 
-        return new EmptySegmentTranslationVersionHandler();
+        return new DummyTranslationVersionHandler();
 
     }
 
@@ -38,34 +40,34 @@ class TranslationVersions extends BaseFeature {
         // status changed or the translation changed
         $user = $params[ 'user' ];
 
-        /** @var \Translations_SegmentTranslationStruct $translation */
+        /** @var Translations_SegmentTranslationStruct $translation */
         $translation = $params[ 'translation' ];
 
-        /** @var \Translations_SegmentTranslationStruct $old_translation */
+        /** @var Translations_SegmentTranslationStruct $old_translation */
         $old_translation = $params[ 'old_translation' ];
 
         $source_page_code = $params[ 'source_page_code' ];
 
-        /** @var \Chunks_ChunkStruct $chunk */
+        /** @var Chunks_ChunkStruct $chunk */
         $chunk = $params[ 'chunk' ];
 
-        /** @var \FeatureSet $features */
+        /** @var FeatureSet $features */
         $features = $params[ 'features' ];
 
-        /** @var \Projects_ProjectStruct $project */
+        /** @var Projects_ProjectStruct $project */
         $project = $params[ 'project' ];
 
-        $sourceEvent = new SegmentTranslationEventModel(
+        $sourceEvent = new TranslationEvent(
                 $old_translation,
                 $translation,
                 $user,
                 $source_page_code
         );
 
-        $batchEventCreator = new BatchEventCreator( $chunk );
-        $batchEventCreator->setFeatureSet( $features );
-        $batchEventCreator->addEventModel( $sourceEvent );
-        $batchEventCreator->setProject( $project );
+        $batchEventHandler = new TranslationEventsHandler( $chunk );
+        $batchEventHandler->setFeatureSet( $features );
+        $batchEventHandler->addEvent( $sourceEvent );
+        $batchEventHandler->setProject( $project );
 
         // If propagated segments exist, start cycle here
         // @TODO COMPLETE REFACTORY IS NEEDED HERE!!!!!
@@ -85,14 +87,14 @@ class TranslationVersions extends BaseFeature {
 
             foreach ( $segmentTranslations as $segmentTranslationBeforeChange ) {
 
-                /** @var \Translations_SegmentTranslationStruct $propagatedSegmentAfterChange */
+                /** @var Translations_SegmentTranslationStruct $propagatedSegmentAfterChange */
                 $propagatedSegmentAfterChange                      = clone $segmentTranslationBeforeChange;
                 $propagatedSegmentAfterChange->translation         = $translation->translation;
                 $propagatedSegmentAfterChange->status              = $translation->status;
                 $propagatedSegmentAfterChange->autopropagated_from = $translation->id_segment;
                 $propagatedSegmentAfterChange->time_to_edit        = 0;
 
-                $propagatedEvent = new SegmentTranslationEventModel(
+                $propagatedEvent = new TranslationEvent(
                         $segmentTranslationBeforeChange,
                         $propagatedSegmentAfterChange,
                         $user,
@@ -100,12 +102,12 @@ class TranslationVersions extends BaseFeature {
                 );
 
                 $propagatedEvent->setPropagationSource( false );
-                $batchEventCreator->addEventModel( $propagatedEvent );
+                $batchEventHandler->addEvent( $propagatedEvent );
             }
         }
 
         try {
-            $batchEventCreator->save();
+            $batchEventHandler->save();
             // $event->setChunkReviewsList( $chunkReviews ) ;
             ( new Jobs_JobDao() )->destroyCacheByProjectId( $chunk->id_project );
             Projects_ProjectDao::destroyCacheById( $chunk->id_project );
